refactor: Apply buffer for all services
