+++ core/benches/oio/write.rs
@@ -17,8 +17,7 @@
 
 use criterion::Criterion;
 use once_cell::sync::Lazy;
-use opendal::raw::oio::AtLeastBufWriter;
-use opendal::raw::oio::ExactBufWriter;
+use opendal::raw::oio::BoundedBufWriter;
 use opendal::raw::oio::Write;
 use rand::thread_rng;
 use size::Size;
@@ -28,33 +27,7 @@ use super::utils::*;
 pub static TOKIO: Lazy<tokio::runtime::Runtime> =
     Lazy::new(|| tokio::runtime::Runtime::new().expect("build tokio runtime"));
 
-pub fn bench_at_least_buf_write(c: &mut Criterion) {
-    let mut group = c.benchmark_group("at_least_buf_write");
-
-    let mut rng = thread_rng();
-
-    for size in [
-        Size::from_kibibytes(4),
-        Size::from_kibibytes(256),
-        Size::from_mebibytes(4),
-        Size::from_mebibytes(16),
-    ] {
-        let content = gen_bytes(&mut rng, size.bytes() as usize);
-
-        group.throughput(criterion::Throughput::Bytes(size.bytes() as u64));
-        group.bench_with_input(size.to_string(), &content, |b, content| {
-            b.to_async(&*TOKIO).iter(|| async {
-                let mut w = AtLeastBufWriter::new(BlackHoleWriter, 256 * 1024);
-                w.write(content.clone()).await.unwrap();
-                w.close().await.unwrap();
-            })
-        });
-    }
-
-    group.finish()
-}
-
-pub fn bench_exact_buf_write(c: &mut Criterion) {
+pub fn bench_bounded_buf_write(c: &mut Criterion) {
     let mut group = c.benchmark_group("exact_buf_write");
 
     let mut rng = thread_rng();
@@ -70,7 +43,7 @@ pub fn bench_exact_buf_write(c: &mut Criterion) {
         group.throughput(criterion::Throughput::Bytes(size.bytes() as u64));
         group.bench_with_input(size.to_string(), &content, |b, content| {
             b.to_async(&*TOKIO).iter(|| async {
-                let mut w = ExactBufWriter::new(BlackHoleWriter, 256 * 1024);
+                let mut w = BoundedBufWriter::new(BlackHoleWriter, 256 * 1024);
                 w.write(content.clone()).await.unwrap();
                 w.close().await.unwrap();
             })
