+++ core/src/raw/oio/write/mod.rs
@@ -39,8 +39,5 @@ mod one_shot_write;
 pub use one_shot_write::OneShotWrite;
 pub use one_shot_write::OneShotWriter;
 
-mod at_least_buf_write;
-pub use at_least_buf_write::AtLeastBufWriter;
-
-mod exact_buf_write;
-pub use exact_buf_write::ExactBufWriter;
+mod bounded_buf_write;
+pub use bounded_buf_write::BoundedBufWriter;
