+++ core/src/services/oss/backend.rs
@@ -15,7 +15,6 @@
 // specific language governing permissions and limitations
 // under the License.
 
-use std::cmp::max;
 use std::collections::HashMap;
 use std::collections::HashSet;
 use std::fmt::Debug;
@@ -39,6 +38,9 @@ use crate::raw::*;
 use crate::services::oss::writer::OssWriters;
 use crate::*;
 
+#[allow(dead_code)]
+/// FIXME: we should use this const when capability has been added.
+///
 /// The minimum multipart size of OSS is 100 KiB.
 ///
 /// ref: <https://www.alibabacloud.com/help/en/oss/user-guide/multipart-upload-12>
@@ -381,7 +383,7 @@ pub struct OssBackend {
 impl Accessor for OssBackend {
     type Reader = IncomingAsyncBody;
     type BlockingReader = ();
-    type Writer = oio::TwoWaysWriter<OssWriters, oio::AtLeastBufWriter<OssWriters>>;
+    type Writer = OssWriters;
     type BlockingWriter = ();
     type Pager = OssPager;
     type BlockingPager = ();
@@ -479,17 +481,6 @@ impl Accessor for OssBackend {
             OssWriters::One(oio::OneShotWriter::new(writer))
         } else {
             OssWriters::Two(oio::MultipartUploadWriter::new(writer))
-        };
-
-        let w = if let Some(buffer_size) = args.buffer_size() {
-            let buffer_size = max(MINIMUM_MULTIPART_SIZE, buffer_size);
-
-            let w =
-                oio::AtLeastBufWriter::new(w, buffer_size).with_total_size(args.content_length());
-
-            oio::TwoWaysWriter::Two(w)
-        } else {
-            oio::TwoWaysWriter::One(w)
         };
 
         Ok((RpWrite::default(), w))
